feat(honker): added comments and failed collection
