+++ packages/queue-honker/src/provider.ts
@@ -1,3 +1,9 @@
+// Honker-over-SQLite implementation of `IQueueProvider`. Registers itself
+// as the "honker" provider at module load; the server picks it up via
+// `import "@bb/queue-honker"` (side effect) + `connectQueue("honker")`.
+//
+// Single-process, single SQLite file at `Config.QueueDbPath`. No network.
+
 import { open, type Database, type Job, type JsonValue, type Queue } from "@russellthehippo/honker-node";
 import { JobType, type JobMessage, type PayloadFor } from "@bb/types";
 import { QueueConnectError, QueueNotConnectedError } from "@bb/errors";
@@ -14,10 +20,17 @@ import type {
 } from "@bb/queue-core";
 import { mapHonkerPriority } from "./priority.ts";
 import { resolveQueueDbPath } from "./paths.ts";
+import { normalizeFailed } from "./failed.ts";
 
+// Visibility = how long Honker waits before reassigning a claimed-but-not-
+// heartbeated job to another worker. 5 min is large enough that GC pauses
+// don't trigger spurious reclaims; heartbeating every 60 s gives us two
+// missed-heartbeat tolerance.
 const VISIBILITY_S = 300;
 const HEARTBEAT_EXTEND_S = 300;
 const HEARTBEAT_MS = 60_000;
+// Match BullMQ's fixed-5s backoff. Without an explicit retry delay, a failed
+// job would wait the full VISIBILITY_S (300s) before being re-claimed.
 const RETRY_DELAY_S = 5;
 const MAX_ATTEMPTS = 3;
 const SWEEP_INTERVAL_MS = 30_000;
@@ -43,11 +56,15 @@ class HonkerQueueProvider implements IQueueProvider {
   async connect(): Promise<void> {
     try {
       const dbPath = resolveQueueDbPath();
+      // Honker auto-creates `_honker_live` / `_honker_dead` / etc. on first
+      // open — no explicit `bootstrap()` call needed.
       const db = open(dbPath);
       for (const type of ALL_JOB_TYPES) {
         this.queues.set(type, db.queue(type, { visibilityTimeoutS: VISIBILITY_S, maxAttempts: MAX_ATTEMPTS }));
       }
       this.db = db;
+      // Periodic sweeper moves attempts-exhausted rows from `_honker_live`
+      // to `_honker_dead`. `unref()` so it doesn't keep the event loop alive.
       this.sweepTimer = setInterval(() => this.sweep(), SWEEP_INTERVAL_MS);
       this.sweepTimer.unref();
     } catch (cause: unknown) {
@@ -62,6 +79,9 @@ class HonkerQueueProvider implements IQueueProvider {
       clearInterval(this.sweepTimer);
       this.sweepTimer = null;
     }
+    // Abort all worker `claim()` iterators, then wait for their `for await`
+    // loops to unwind before closing the database (so no read happens
+    // against a closed handle).
     const loops = this.workers.splice(0);
     for (const w of loops) {
       w.controller.abort();
@@ -91,6 +111,10 @@ class HonkerQueueProvider implements IQueueProvider {
   ): Promise<string> {
     const db = this.requireDb();
     const queue = this.requireQueue(type);
+    // Honker has no native stable-jobId concept (BullMQ's dedupe primitive),
+    // so we emulate it by querying `_honker_live` for an existing row with
+    // matching knowledgeId. If found, return its id — re-publishing the
+    // same logical job is a no-op (per the IQueueProvider contract).
     const existing = db.query(
       "SELECT id FROM _honker_live WHERE queue = ? AND json_extract(payload, '$.knowledgeId') = ? LIMIT 1",
       [type, message.knowledgeId],
@@ -108,6 +132,9 @@ class HonkerQueueProvider implements IQueueProvider {
   registerWorker<T extends JobType>(type: T, handler: JobHandler<T>, opts: WorkerRegistrationOptions = {}): void {
     const queue = this.requireQueue(type);
     const concurrency = opts.concurrency ?? defaultConcurrencyFor(type);
+    // N independent loops with batch_size=1 each — prevents head-of-line
+    // blocking from a slow job in the same claim batch. Matches BullMQ's
+    // `concurrency` semantics (N jobs in flight per type).
     for (let i = 0; i < concurrency; i++) {
       const workerId = `${type}-${i}-${process.pid}`;
       const controller = new AbortController();
@@ -123,10 +150,14 @@ class HonkerQueueProvider implements IQueueProvider {
     signal: AbortSignal,
   ): Promise<void> {
     try {
+      // `queue.claim()` is an AsyncIterableIterator that wakes on DB updates
+      // or due deadlines (no polling). The signal lets `close()` end the
+      // iterator cleanly.
       for await (const job of queue.claim(workerId, { signal })) {
         await this.processJob(job, workerId, handler);
       }
     } catch (err) {
+      // Aborts during shutdown are expected — log only on real crashes.
       if (signal.aborted) {
         return;
       }
@@ -135,6 +166,12 @@ class HonkerQueueProvider implements IQueueProvider {
   }
 
   private async processJob<T extends JobType>(job: Job, workerId: string, handler: JobHandler<T>): Promise<void> {
+    // Lease ownership tracking. If `heartbeat()` ever returns false the
+    // visibility window expired (event loop stall, long sync work) and
+    // another worker has reclaimed this job. Policy: stop-and-finish —
+    // log it, stop heartbeating, let the handler complete normally.
+    // Both workers' writes converge at the destination via idempotent
+    // upserts; the cost is duplicate LLM tokens on the rare event.
     let ownsLease = true;
     const hb = setInterval(() => {
       if (!ownsLease) {
@@ -148,12 +185,18 @@ class HonkerQueueProvider implements IQueueProvider {
     }, HEARTBEAT_MS);
     try {
       await handler(job.payload as unknown as JobMessage<PayloadFor<T>>);
+      // Only ack if we still own the lease — otherwise the row already
+      // belongs to another worker; double-acking is a no-op but it's
+      // cleaner not to issue the write.
       if (ownsLease) {
         job.ack();
       }
     } catch (err: unknown) {
       const reason = describeError(err);
       logger.error(`queue-honker: job=${job.id} handler threw: ${reason}; scheduling retry`);
+      // `retry(delayS, error)` increments `attempts` and re-queues with a
+      // 5s delay (matches BullMQ parity). After MAX_ATTEMPTS exhaustions,
+      // the next `sweepExpired()` tick moves the row to `_honker_dead`.
       if (ownsLease) {
         job.retry(RETRY_DELAY_S, reason);
       }
@@ -165,6 +208,11 @@ class HonkerQueueProvider implements IQueueProvider {
   async removeKnowledgeJobs(knowledgeId: string): Promise<RemoveKnowledgeJobsResult> {
     const db = this.requireDb();
     const placeholders = ALL_JOB_TYPES.map(() => "?").join(",");
+    // Best-effort delete across all queues by knowledgeId. Honker's Queue
+    // API doesn't expose a `cancel` method — direct table mutation is the
+    // documented escape hatch (per the user-supplied SQL reference).
+    // Wrapped in a transaction so partial failures don't leave half-deleted
+    // state visible to a concurrent claim.
     const tx = db.transaction();
     try {
       const removed = tx.execute(
@@ -189,6 +237,7 @@ class HonkerQueueProvider implements IQueueProvider {
     return rows.map(normalizeFailed);
   }
 
+  /** Move attempts-exhausted rows from `_honker_live` to `_honker_dead`. */
   private sweep(): void {
     for (const queue of this.queues.values()) {
       try {
@@ -215,41 +264,8 @@ class HonkerQueueProvider implements IQueueProvider {
   }
 }
 
-function normalizeFailed(row: Record<string, unknown>): FailedJob {
-  const payloadRaw = row["payload"];
-  const payload: unknown = typeof payloadRaw === "string" ? safeParseJson(payloadRaw) : payloadRaw;
-  const knowledgeId = extractKnowledgeId(payload);
-  const queue = typeof row["queue"] === "string" ? row["queue"] : "";
-  const diedAt = typeof row["died_at"] === "number" ? row["died_at"] : Date.now() / 1000;
-  return {
-    id: String(row["id"] ?? ""),
-    type: queue as JobType,
-    knowledgeId,
-    attempts: typeof row["attempts"] === "number" ? row["attempts"] : 0,
-    failedAt: new Date(diedAt * 1000).toISOString(),
-    reason: typeof row["last_error"] === "string" ? row["last_error"] : "",
-    payload,
-  };
-}
-
 function describeError(err: unknown): string {
   return err instanceof Error ? err.message : String(err);
-}
-
-function safeParseJson(s: string): unknown {
-  try {
-    return JSON.parse(s);
-  } catch {
-    return null;
-  }
-}
-
-function extractKnowledgeId(payload: unknown): string {
-  if (payload !== null && typeof payload === "object" && "knowledgeId" in payload) {
-    const v = (payload as { knowledgeId: unknown }).knowledgeId;
-    return typeof v === "string" ? v : "";
-  }
-  return "";
 }
 
 registerQueueProvider("honker", () => new HonkerQueueProvider());
