+++ packages/queue/src/registry.ts
@@ -1,7 +1,17 @@
+// Provider registry — the facade core. Mirrors the `@bb/db` / `@bb/graph-db`
+// pattern: side-effect imports of provider packages (`@bb/queue-bullmq`,
+// `@bb/queue-honker`) call `registerQueueProvider(name, factory)` at module
+// load. The server then picks one with `connectQueue(Config.QueueProvider)`.
+//
+// Switching providers is a cold cutover — only one is active at a time;
+// `closeQueue()` must be called before re-connecting under a different name.
+
 import { QueueNotConnectedError } from "@bb/errors";
 import type { IQueueProvider, QueuePingResult } from "@bb/queue-core";
 
 let activeProvider: IQueueProvider | null = null;
+// In-flight connect promise: dedupes concurrent `connectQueue` callers so
+// two parallel callers don't construct two providers and race.
 let connecting: Promise<void> | null = null;
 const providers = new Map<string, () => IQueueProvider>();
 
@@ -36,6 +46,9 @@ async function doConnect(providerName: string): Promise<void> {
   }
   const next = factory();
   await next.connect();
+  // Only set `activeProvider` after `connect()` succeeds — if it throws the
+  // facade stays in the "not connected" state and `connectQueue` can be
+  // retried cleanly.
   activeProvider = next;
 }
 
@@ -43,6 +56,8 @@ export async function closeQueue(): Promise<void> {
   if (activeProvider === null) {
     return;
   }
+  // Clear the reference *before* awaiting `close()` so a concurrent
+  // `connectQueue` call doesn't see a half-torn-down provider.
   const c = activeProvider;
   activeProvider = null;
   await c.close();
